p4a_requirements.txt
python3=3.11.5
kivy
sqlite3
mysql-connector-python